Create ModelCallHexa2.py
The RF model was saved with the help of the pickle module, which defines a technique for converting any Python object to a sequence of bytes. This is often referred to as "serializing" the item. The object's byte stream may be transferred or saved and rebuilt to generate a new object with the same properties. The ANN model's features (obtained structure, weight, and bias values) were recorded using the save function on the model and giving a filename, which resulted in the model's information being saved as a JSON file. The JSON file will allow the designer to use the stepwise algorithm to predict the crashworthiness parameters. This step can be done by simulating all possible combinations of crashworthiness parameters and averaging down the corresponding results. For instance, testing was done at 60o and 65o hexagonal angles; however, the generated models can estimate the crashworthiness characteristics for a 63o hexagonal angle using the trained model.